requirements.txt
asttokens==3.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==41.0.7
decorator==5.2.1
executing==2.2.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
pan-os-python==1.12.2
pan-python==0.17.0
pandevice==0.14.0
panos-upgrade-assurance==1.1.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==23.3.0
requests==2.32.3
setuptools==80.4.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.4.0
wcwidth==0.2.13
xmltodict==0.14.2
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
future==1.0.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.5.0
ntc_templates==7.8.0
pan-os-python==1.12.2
pan-python==0.17.0
paramiko==3.5.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
setuptools==80.4.0
six==1.17.0
tabulate==0.9.0
textfsm==1.1.3
urllib3==2.4.0
